hx711.py: optimization of code & some enhancements
1. Optimization of the read() method.

The core loop is optimized and the data bits are directly
collected into the result value. And interrupts are disabled
during the high phase of the clock pulse. That ensures that
the positive clock pulse is never longer than 50 µs.
Then, all code related to the alltrue state is obsolete, and
was removed.

2. Adding a lowpass filtered read method.

That consists of the methods read_lowpass() and set_time_constant().
read_lowpass() is somewhat comparable to a sliding average, but returns
a value at the rate of the underlying read() call. The time constant
for the filter is set with set_time_constant. The range is 0 to 1.
1 is equivalent to no filter at all, 0 means inifinite time constant.
Reasonable values are 0.01 to 0.75.