hx711_spi.py, hxtest.py: A variant which uses the SPI bus; and test code
This variant is fucntionally identical to the pin toggle version, but
uses the SPI bus. MOSI is conectesd to the HX711 clock input, and MISO is
connected to the HX711 data out. That ensure a reliable timing for the HX711.
The code is slightly different.
hxtest.py is a small test code which collects the distribution of values
over time. It also take sthe temperature. I made it to determine the
variantion of the test data ovet time an temperature.
To be done: README.md